% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_summed_exps.R
\name{log_summed_exp}
\alias{log_summed_exp}
\title{Computes the log of summed exponents.}
\usage{
log_summed_exp(vec)
}
\arguments{
\item{vec}{A vector of type integer or double, representing the values of the exponentials.}
}
\value{
A number representing the log of the sum of the exponentials.
}
\description{
Calculates the expression of the form:
log( sum_{i=1}^n  exp(x_i))
From a vector of values (x_1,...,x_i,...,x_n)
}
\examples{
log_summed_exp(c(1,2,3,4,5))
vec <- c(1:10)
log_summed_exp(vec)
}
\author{
Robert Trujillo
}
